Added middlewares for Tasks

+++ routes/web.php
@@ -21,19 +21,19 @@ Auth::routes();
 
     Route::get('/user/tasks','UserController@showtasks')->name('User.tasks');
 
-    Route::post('/task','TaskController@store')->name('Task.store');
-    Route::get('/Task/{id}', 'TaskController@show')->name('Task.show');
-    Route::get('/task/create','TaskController@create')->name('Task.create');
-    Route::get('/task/{id}/edit','TaskController@edit')->name('Task.edit');
-    Route::patch('/task/update', 'TaskController@update')->name('Task.update');
-    Route::delete('/task/{id}/delete', 'TaskController@destroy')->name('Task.destroy');
+    Route::post('/task','TaskController@store')->name('Task.store')->middleware('isRelatedtoTask');
+    Route::get('/Task/{id}', 'TaskController@show')->name('Task.show')->middleware('isRelatedtoTask');
+    Route::get('/task/create','TaskController@create')->name('Task.create')->middleware('isRelatedtoTask');
+    Route::get('/task/{id}/edit','TaskController@edit')->name('Task.edit')->middleware('isRelatedtoTask')->middleware('isAdmin');
+    Route::patch('/task/update', 'TaskController@update')->name('Task.update')->middleware('isRelatedtoTask')->middleware('isAdmin');
+    Route::delete('/task/{id}/delete', 'TaskController@destroy')->name('Task.destroy')->middleware('isRelatedtoTask')->middleware('isAdmin');
 
-    Route::get('/task/{id}/viewers','TaskController@viewers')->name('Task.viewers');
-    Route::get('/task/{id}/desviewer/{user_id}','TaskController@desviewer')->name('viewer.destroy');
-    Route::POST('/task_viewer','TaskController@addviewer')->name('viewer.add');
-    Route::POST('/task/edviro','TaskController@editviewerrole')->name('viewer.editrole');
+    Route::get('/task/{id}/viewers','TaskController@viewers')->name('Task.viewers')->middleware('isRelatedtoTask');
+    Route::get('/task/{id}/desviewer/{user_id}','TaskController@desviewer')->name('viewer.destroy')->middleware('isRelatedtoTask')->middleware('isAdmin');
+    Route::POST('/task_viewer','TaskController@addviewer')->name('viewer.add')->middleware('isRelatedtoTask')->middleware('isAdmin');
+    Route::POST('/task/edviro','TaskController@editviewerrole')->name('viewer.editrole')->middleware('isRelatedtoTask')->middleware('isAdmin');
 
-    Route::get('/task/{id}/subtasks','TaskController@showsubtasks')->name('Task.subtasks');
+    Route::get('/task/{id}/subtasks','TaskController@showsubtasks')->name('Task.subtasks')->middleware('isRelatedtoTask');
 
     Route::post('/subtask','SubtaskController@store')->name('Subtask.store');
     Route::get('/subtask/{id}', 'SubtaskController@show')->name('Subtask.show');
